Changes in send value and monitor, refactor quickstart into multiple components
