+++ docusaurus/src/pages/quickstart-steps/index.js
@@ -1,0 +1,5 @@
+export * from './GettingStarted';
+export * from './GenerateWallet';
+export * from './FundWallet';
+export * from './SendValue';
+export * from './StoreData';
